+++ src/formatter.rs
@@ -295,6 +295,7 @@ pub mod tests {
             Some("test@example.com".to_string()),
             subject.as_ref(),
             message.as_ref().to_string(),
+            "".to_string(),
             vec!["src/main.rs".to_string()],
         )
     }