+++ src/rules/message_skip_build_tag.rs
@@ -101,6 +101,7 @@ mod tests {
             Some("test@example.com".to_string()),
             "Some subject",
             "Some message".to_string(),
+            "".to_string(),
             files,
         )
     }